Started on recombine args method and found bug
Started working on createProcess, Realized it takes a whole commandline. But because i want to handle built in commands as well it is more efficient to simply recombine my already parsed args. Also found a bug where my shiftTerminalStringRight method seems to have a memory leak if there are two or more blankspaces to the right of my cursor